+++ squads/tasks.py
@@ -1,5 +1,1 @@
 """App Tasks"""
-
-from squads.hooks import get_extension_logger
-
-logger = get_extension_logger(__name__)