+++ squads/tests/test_groups.py
@@ -5,13 +5,15 @@ from squads.forms import SquadsGroupForm
 
 class SquadsGroupFormTest(TestCase):
     def test_clean_description_too_long(self):
-        # Erstellen Sie Formulardaten mit einer Beschreibung, die länger als 1000 Zeichen ist
+        # given
         form_data = {
             "name": "Test Group",
-            "description": "a" * 1001,  # 1001 Zeichen lange Beschreibung
+            "description": "a" * 1001,
             "require_approval": False,
         }
+        # when
         form = SquadsGroupForm(data=form_data)
+        # then
         self.assertFalse(form.is_valid())
         self.assertIn("description", form.errors)
         self.assertEqual(
@@ -20,22 +22,27 @@ class SquadsGroupFormTest(TestCase):
         )
 
     def test_clean_description_valid(self):
-        # Erstellen Sie Formulardaten mit einer gültigen Beschreibung
+        # given
         form_data = {
             "name": "Test Group",
             "description": "<p>This is a valid description.</p>",
             "require_approval": False,
         }
+        # when
         form = SquadsGroupForm(data=form_data)
+        # then
         self.assertTrue(form.is_valid())
 
     def test_clean_description_invalid_tags(self):
+        # given
         form_data = {
             "name": "Test Group",
             "description": '<script>alert("Hack!");</script><p>This is a valid part of the description.</p>',
             "require_approval": False,
         }
+        # when
         form = SquadsGroupForm(data=form_data)
+        # then
         self.assertTrue(form.is_valid())
         cleaned_description = form.cleaned_data["description"]
         self.assertNotIn("<script>", cleaned_description)